Move <TypeName>/ErrorMessage to <TypeName>/internal/ErrorMessage
