+++ packages/option-t/src/Maybe/mapAsync.ts
@@ -8,7 +8,7 @@ import {
     type NotNullOrUndefined,
     expectNotNullOrUndefined,
 } from './Maybe.js';
-import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_MAYBE } from './ErrorMessage.js';
+import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_MAYBE } from './internal/ErrorMessage.js';
 
 function check<T>(value: Maybe<T>): T {
     const passed = expectNotNullOrUndefined(
