+++ packages/option-t/src/Maybe/mapOr.ts
@@ -2,7 +2,7 @@ import { type Maybe, type NotNullOrUndefined, expectNotNullOrUndefined } from '.
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_MAYBE,
     ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_MAYBE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import type { TransformFn } from '../internal/Function.js';
 
 /**
