+++ packages/option-t/src/Maybe/mapOrAsync.ts
@@ -5,7 +5,7 @@ import type { AsyncTransformFn } from '../internal/Function.js';
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_MAYBE,
     ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_MAYBE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import {
     type Maybe,
     isNullOrUndefined,
