+++ packages/option-t/src/Nullable/map.ts
@@ -1,5 +1,5 @@
 import { expectNotNull } from './expect.js';
-import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './ErrorMessage.js';
+import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './internal/ErrorMessage.js';
 import type { TransformFn } from '../internal/Function.js';
 import { type Nullable, isNull, type NotNull } from './Nullable.js';
