+++ packages/option-t/src/Nullable/mapOr.ts
@@ -3,7 +3,7 @@ import { expectNotNull } from './expect.js';
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
     ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_NULLABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import type { TransformFn } from '../internal/Function.js';
 
 /**
