+++ packages/option-t/src/Nullable/mapOrAsync.ts
@@ -5,7 +5,7 @@ import type { Nullable } from './Nullable.js';
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
     ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_NULLABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import { expectNotNull } from './expect.js';
 
 function check<T>(value: Nullable<T>): T {
