+++ packages/option-t/src/Nullable/mapOrElse.ts
@@ -2,7 +2,7 @@ import { expectNotNull } from './expect.js';
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
     ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import type { TransformFn, RecoveryFn } from '../internal/Function.js';
 import { isNotNull, type NotNull, type Nullable } from './Nullable.js';
