+++ packages/option-t/src/Nullable/mapOrElseAsync.ts
@@ -8,7 +8,7 @@ import {
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
     ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import { expectNotNull } from './expect.js';
 
 /**
