+++ packages/option-t/src/Nullable/unwrapOrElse.ts
@@ -1,7 +1,7 @@
 import type { RecoveryFn } from '../internal/Function.js';
 import { isNotNull, type NotNull, type Nullable } from './Nullable.js';
 import { expectNotNull } from './expect.js';
-import { ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './ErrorMessage.js';
+import { ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './internal/ErrorMessage.js';
 
 /**
  *  Return _input_ as `T` if the passed _input_ is not `null`.
