+++ packages/option-t/src/Nullable/unwrapOrElseAsync.ts
@@ -3,7 +3,7 @@ import { ERR_MSG_RECOVERER_MUST_RETURN_PROMISE } from '../internal/ErrorMessage.
 import type { AsyncRecoveryFn } from '../internal/Function.js';
 import { isNotNull, type NotNull, type Nullable } from './Nullable.js';
 import { expectNotNull } from './expect.js';
-import { ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './ErrorMessage.js';
+import { ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE } from './internal/ErrorMessage.js';
 
 function check<T>(value: Nullable<T>): T {
     const passed = expectNotNull(value, ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_NULLABLE);
