+++ packages/option-t/src/Undefinable/map.ts
@@ -1,4 +1,4 @@
-import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE } from './ErrorMessage.js';
+import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE } from './internal/ErrorMessage.js';
 import { expectNotUndefined } from './expect.js';
 import type { TransformFn } from '../internal/Function.js';
 import { type Undefinable, isUndefined, type NotUndefined } from './Undefinable.js';
