+++ packages/option-t/src/Undefinable/mapAsync.ts
@@ -1,6 +1,6 @@
 import type { AsyncTransformFn } from '../internal/Function.js';
 import { type Undefinable, isUndefined, type NotUndefined } from './Undefinable.js';
-import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE } from './ErrorMessage.js';
+import { ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE } from './internal/ErrorMessage.js';
 import { expectNotUndefined } from './expect.js';
 import { assertIsPromise } from '../internal/assert.js';
 import { ERR_MSG_TRANSFORMER_MUST_RETURN_PROMISE } from '../internal/ErrorMessage.js';
