+++ packages/option-t/src/Undefinable/mapOrAsync.ts
@@ -6,7 +6,7 @@ import { type Undefinable, isUndefined, type NotUndefined } from './Undefinable.
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE,
     ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_UNDEFINABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import { expectNotUndefined } from './expect.js';
 
 function check<T>(value: Undefinable<T>): T {
