+++ packages/option-t/src/Undefinable/mapOrElseAsync.ts
@@ -8,7 +8,7 @@ import {
 import {
     ERR_MSG_TRANSFORMER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE,
     ERR_MSG_RECOVERER_MUST_NOT_RETURN_NO_VAL_FOR_UNDEFINABLE,
-} from './ErrorMessage.js';
+} from './internal/ErrorMessage.js';
 import { expectNotUndefined } from './expect.js';
 
 /**
