+++ packages/option-t/src/Undefinable/unwrapOr.ts
@@ -1,6 +1,6 @@
 import { isNotUndefined, type NotUndefined, type Undefinable } from './Undefinable.js';
 import { expectNotUndefined } from './expect.js';
-import { ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_UNDEFINABLE } from './ErrorMessage.js';
+import { ERR_MSG_DEFAULT_VALUE_MUST_NOT_BE_NO_VAL_FOR_UNDEFINABLE } from './internal/ErrorMessage.js';
 
 /**
  *  Return _input_ as `T` if the passed _input_ is not `undefined`.
